correçoes de bug back e front

+++ carvao-app.Models/Requests/GerarReciboRequest.cs
@@ -10,7 +10,7 @@ namespace carvao_app.Models.Requests
     public class GerarReciboRequest
     {
         public int pedido_id { get; set; }
-        public int valor_pago { get; set; }
+        public decimal valor_pago { get; set; }
         public int forma_pagamento { get; set; }
         public string nome_pagador{ get; set; }
         public string observacoes { get; set; }
